Move tom's test to new peoject
